Вернул Methods
