+++ English-Bot.sln
@@ -3,7 +3,12 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29411.108
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "English-Bot", "English-Bot\English-Bot.csproj", "{9DE3BE4F-31D9-4C73-B7C4-E1169AF304CF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "English-Bot", "English-Bot\English-Bot.csproj", "{9DE3BE4F-31D9-4C73-B7C4-E1169AF304CF}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5EA48A39-BA8E-4BDB-AD56-C1D74DC5E38B}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
