+++ English-Bot/Dictionary.cs
@@ -6,11 +6,24 @@ namespace English_Bot
     public class Dictionary
     {
         private Dictionary<long, Word> dict;
-      
+        private Dictionary<string, long[]> eng_ids; 
+        private Dictionary<string, long[]> rus_ids; 
         
         public Dictionary()
         {
             dict = new Dictionary<long, Word>();
+            foreach (var word in Methods.Deserialization<Word>(@"Json dicts/eng_words.json"))
+            {
+                dict.Add(word.id, word);
+                if (eng_ids.ContainsKey(word.eng))
+                    eng_ids[word.eng].Add(word.id);
+                else 
+                    eng_ids.Add(word.eng, word.id);
+                if (rus_ids.ContainsKey(word.rus))
+                    rus_ids[word.rus].Add(word.id);
+                else
+                    rus_ids.Add(word.rus, word.id);
+            }
         }
         /// <summary>
         /// индексация с 1
@@ -30,42 +43,21 @@ namespace English_Bot
                 return dict[id];
             else
                 return null;
+        }
 
+        /// <summary>
+        /// получаем id английских слова  по нашему запросу  
+        /// </summary>
+        public long[] GetWordEng(string word)
+        {
+            return eng_ids[word];
         }
         /// <summary>
-        /// получаем английские слова  по нашему запросу  
+        /// получаем id русских слова  по нашему запросу  
         /// </summary>
-        public Word[] GetWordEng(string word)
+        public long[] GetWordRus(string word)
         {
-            //работает ,но вроде как такие запросы выполняются медленнее чем обычный цикл
-            // var lst = dict.Select(t=>t.Value).Where(d => d.eng.Contains(word));
-
-            List<Word> lst = new List<Word>();
-            
-            foreach (var t in dict)
-            {
-                if (t.Value.eng.Contains(word.ToLower()))
-                    lst.Add(t.Value);
-            }
-            
-            return lst.ToArray();
-        }
-        /// <summary>
-        /// получаем русские слова  по нашему запросу  
-        /// </summary>
-        public Word[] GetWordRus(string word)
-        {
-
-            ////работает ,но вроде как такие запросы выполняются медленнее чем обычный цикл
-            //var lst = dict.Select(t => t.Value).Where(d => d.rus.Contains(word));
-            List<Word> lst = new List<Word>();
-
-            foreach (var t in dict)
-            {
-                if (t.Value.rus.Contains(word.ToLower()))
-                    lst.Add(t.Value);
-            }
-            return lst.ToArray();
+            return rus_ids[word];
         }
 
         public bool AddWord(Word w)
